adding gdzs by admin(me)

+++ bot/handlers/user.py
@@ -3,13 +3,18 @@ from email.message import Message
 
 from aiogram import F, Router, types
 from aiogram.fsm.context import FSMContext
+from aiogram.utils.markdown import hide_link
+
 from bot.classes.classes import UserInfo, loginMsg
 from bot.database import main
-from bot.functions import (gdzAsking, info, lessonsToday, recognize, shedule,
-                           tommorow, tommorowHw, hwSend, updatee)
+from bot.database.main import BooksDB
+from bot.functions import (gdzAsking, hwSend, info, lessonsToday, recognize,
+                           shedule, tommorow, tommorowHw)
+from bot.functions.api import idKlass
 from bot.keyboards import reply
 
 router = Router()
+dbm = BooksDB()
 
 @router.message(commands=["editData"])
 async def editReg(msg: Message, state: FSMContext):
@@ -17,62 +22,57 @@ async def editReg(msg: Message, state: FSMContext):
     await state.set_state(UserInfo.QL)
 
 
-# Routes updates telegram bot command 'update' to this function using decorator.
-@router.message(commands=['update'])
-# Asynchronous function declaration with message object as arguement and returning None.
-async def update(message: Message) -> None:
-    # Checking if message is coming from chat with specific chat ID.
-    if message.chat.id == 882076783:
-        # If yes then calling the `updateq()` and sending its output in HTML code format.
-        await message.answer(f'<code>{updatee.updateq()}</code>' if updatee.updateq() else 'no')
-
-    
 @router.message(commands=['tl'])
 async def tlcmd(message: Message) -> None:
-        await shedule.Tlschedule(message)
+    await shedule.Tlschedule(message)
+
 
 @router.message(commands=['ts'])
 async def tscmd(message: Message) -> None:
     """Текущее расписание на день"""
     items = await shedule.days_schedule()
-    strings = [f"{str(key).capitalize()}: {item}" for key, item in items.items()]
+    strings = [f"{str(key).capitalize()}: {item}" for key,
+               item in items.items()]
     result = "\n".join(strings)
     await message.answer(result)
+
 
 @router.message(commands=['bydit'])
 async def bdnycmd(message: Message) -> None:
     await message.answer('Не будет')
 
+
 @router.message(commands=['srs'])
 async def srscmd(message: Message) -> None:
     await message.answer('srs')
 
+
 @router.callback_query(text="nope")
-async def cancel(callback_query, state: FSMContext):
+async def cancel(callback_query):
     await callback_query.message.delete()
+
 
 @router.callback_query(text="edit")
 async def edit(msg: Message, state: FSMContext):
     await msg.answer('пока не воркает')
+
 
 @router.callback_query(text="ok")
 async def ok(callback_query, state: FSMContext):
     await callback_query.answer('пока не воркает')
     mes, pg, nm, sj = await hwSend.hwSend(callback_query.message.html_text)
     r = await callback_query.message.answer(mes)
-    if sj == 'физика':
+    await callback_query.message.delete()
+    url = dbm.get(sj)
+    if url != -1:
+        url = url['url']
+        nonfileurl = url['nonfileurl']
+        donedUrl = eval(f"f'{url}'") #needs to be chaged #TODO
+        await callback_query.message.answer(f'{hide_link(donedUrl)}'
+                                            f'{nonfileurl}') 
+    else:
         await r.delete()
-        await callback_query.message.answer(f'гдз по Физике: https://reshak.ru/reshebniki/fizika/9/perishkin/images1/new/Upr/{int(nm)+4 if int(nm)>20 else nm}.png')
-    elif sj == 'иностранный язык (английский)':
-        await r.delete()
-        await callback_query.message.answer(f'гдз по Англу: https://fotogdz.ru/media/isbns/496/answers/496_-_{pg}.jpg')
-    
-
-# @router.message(commands=["reset"])
-# async def update(msg: Message):
-#     main.DiaryDB('database.db').resetdb()
-#     await msg.answer("Кнопки были успешно обновлены!", reply_markup=reply.start)
-
+        await callback_query.message.answer('Я пока не умею')
 
 @router.message(commands=["start"])
 async def start(msg: types.Message, state: FSMContext):
@@ -82,6 +82,7 @@ async def start(msg: types.Message, state: FSMContext):
         await state.set_state(UserInfo.QL)
     else:
         await msg.answer("Кнопки были успешно обновлены!", reply_markup=reply.start)
+
 
 @router.message(F.text == "домашка завтра")
 async def homeworkTM(msg: Message):
@@ -94,7 +95,7 @@ async def homeworkTM(msg: Message):
 @router.message(F.text == "предметы завтра")
 async def lessonsTommorow(msg: Message):
     r = await msg.answer('<b>Секунду...</b>')
-    text=await tommorow.tommorow(msg)
+    text = await tommorow.tommorow(msg)
     await r.delete()
     await msg.answer(text, reply_markup=reply.start)
 
@@ -102,7 +103,7 @@ async def lessonsTommorow(msg: Message):
 @router.message(F.text == "хто я")
 async def me(msg: Message):
     r = await msg.answer('<b>Секунду...</b>')
-    text=await info.info(msg)
+    text = await info.info(msg)
     await r.delete()
     await msg.answer(text, reply_markup=reply.start)
 
@@ -110,16 +111,20 @@ async def me(msg: Message):
 @router.message(F.text == "уроки сегодня")
 async def lessonesToday(msg: Message):
     r = await msg.answer('<b>Секунду...</b>')
-    text=await lessonsToday.todaylessons(msg)
+    text = await lessonsToday.todaylessons(msg)
     await r.delete()
     await msg.answer(text, reply_markup=reply.start)
 
+
 @router.message(F.text == 'Гдз запрос ура')
 async def gdz(msg: Message):
-    r = await msg.answer('<b>Секунду...</b>')
-    dsadsd, orig= await recognize.recohniz(msg)
-    if dsadsd == -99 and orig == -91:
-        await r.edit_text(text='ошибка, попробуй войти с помощью команды /editData или напиши Денису')
+    if idKlass(msg.chat.id) == '9':
+        r = await msg.answer('<b>Секунду...</b>')
+        dsadsd, orig = await recognize.recohniz(msg)
+        if dsadsd == -99 and orig == -91:
+            await r.edit_text(text='ошибка, попробуй войти с помощью команды /editData или напиши Денису')
+        else:
+            await r.delete()
+            await gdzAsking.asking(msg, dsadsd, orig)
     else:
-        await r.delete()
-        await gdzAsking.asking(msg, dsadsd, orig)
+        await msg.answer('Для твоего класса пока недоступно')
